unifi host update readme
